Update minichat input color if maxsize is reached
This commit adds or remove the has-warning class to the parent of the
minichat input depending if the maximum number of characters has been
reached.

It is using the has-warning because this is not an error; the user can
still send his message, but it will not be able to add more characters
to the message.