Implement websockets
This commit implements websockets for the minichat and the
notifications.

Side effects of this commit:
- Update to Django 1.9
- Improved travis tests
- New selenium tests

Websockets Support
==================

The library used to handle websockets is
[django-websocket-redis][ws4redis].

It has been chosen for the following reasons:
- It's a popular, maintained library
- Uses redis as a backend. Redis is very popular, so we should not have
  too much troubles with it.

We use the library on the server side and the client side.

It means that if you want a full instance of Lexpage running with all
the features, you will need a redis server running. If you don't, you
will not be able to have the refresh of the minichat and the
notifications.

[ws4redis]:http://django-websocket-redis.readthedocs.org/en/latest/usage.html

How we implemented websockets
=============================

On the client side, we use an improved javascript library that
originally came from the django-websocket-redis. The modifications done
are for adding extra callback functions.

On the server side, we use the Django Signals mechanism to send the
messages. It means that it does not matter the place in the code where
we actually change a model, each change will create a websocket event.

We currently use 2 different websockets:

Minichat
--------

As the minichat already implemented some kind of refresh, we decided to
simply reuse that code but trigger it with the websocket, removing the
need to refresh on a regular basis.

When you post a message, the minichat will only be notified by the
websocket. Before, it was refreshed by the chat form itself.

The content of the websocket message does not contain anything useful,
as it is just a trigger for a full reload of the minichat <div>.

That websocket uses the 'broadcast' events, so the events are sent to
anyone.

From the user point of view, when the connection to the websocket is
not active, the textinput for the minichat is disabled. This will help
user to see if the chat is live or not, and hopefully help them to see
when there are troubles with websockets.

Notifications
-------------

The notifications websocket works as the minichat one, with the
difference that two places are updated: one for the desktop view and one
for the mobile view.

That websocket uses the 'subscribe-user' events, so you only get the
notifications for the current user (and you are not informed when other
users' notifications change).

Testing
=======

We are running basic tests with websockets:

1. Posting a message to the minichat
2. Posting and correcting a message to the minichat (using s/foo/bar)
3. Creating a minichat message with the Django ORM (Message().save())
   and checking that it appears on the minichat.
4. Posting a message which highlight the current user, in that way it
   creates a notification. Then we check that the number of notifications
   has increased.
5. Create notifications and dismiss them without the need to re-open the
   dropdown.

To make these tests work, it was needed to set up Selenium, with a
database that supports locking. To be closer to production, we have
chosen postgresql, which is available on travis.

We updated to Django 1.9 because of the following commit that is
available only in that version and that allowed us to rewrite only a
small part of the testing classes:
https://github.com/django/django/commit/a7901c2e090d720ef0427552387be7f14cbe2ac6

We needed a custom implementation of the DjangoWebserver that would
support multiple connections. This is done in the file
`app/tests_helpers.py`, using the new django 1.9 functions.

The travis tests are running with the latest version of Firefox and
postgresql 9.4.

Bugs closed:
- Fix #42
- Fix #54